requirements.txt
streamlit==1.28.0
pypdf
sentence-transformers==2.2.2
transformers==4.33.2
faiss-cpu
torch==2.2.0
